pyproject.toml
[project]
name = "maccin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "asyncio==4.0.0",
    "bcrypt==5.0.0",
    "cffi==2.0.0",
    "contourpy==1.3.3",
    "cryptography==46.0.3",
    "cycler==0.12.1",
    "ecdsa==0.19.1",
    "et-xmlfile==2.0.0",
    "filelock==3.20.0",
    "fonttools==4.60.1",
    "fsspec==2025.10.0",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "kiwisolver==1.4.9",
    "markupsafe==3.0.3",
    "matplotlib==3.10.7",
    "mpmath==1.3.0",
    "networkx==3.5",
    "numpy==2.3.4",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pandas==2.3.3",
    "passlib==1.7.4",
    "pillow==12.0.0",
    "pyasn1==0.6.1",
    "pycparser==2.23",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.11.0",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-jose==3.5.0",
    "pytz==2025.2",
    "redis==7.1.0",
    "rsa==4.9.1",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "seaborn==0.13.2",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "torch==2.9.0",
    "torchvision==0.24.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "websockets==15.0.1",
]
